% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plot.R
\name{plot.qgcompemmfit}
\alias{plot.qgcompemmfit}
\title{Default plotting method for a qgcompfit object}
\usage{
\method{plot}{qgcompemmfit}(x, emmval = 0, suppressprint = FALSE, ...)
}
\arguments{
\item{x}{"qgcompfit" object from \code{qgcomp.noboot},  \code{qgcomp.boot},
\code{qgcomp.cox.noboot},  \code{qgcomp.cox.boot}, \code{qgcomp.zi.noboot} or \code{qgcomp.zi.boot} functions}

\item{emmval}{fixed value for effect measure modifier at which pointwise comparisons are calculated}

\item{suppressprint}{If TRUE, suppresses the plot, rather than printing it
by default (it can be saved as a ggplot2 object (or list of ggplot2 objects if x is from a zero-
inflated model) and used programmatically)
(default = FALSE)}

\item{...}{unused}
}
\description{
Plot a quantile g-computation object. For qgcomp.noboot, this function will
create a butterfly plot of weights. For qgcomp.boot, this function will create
a box plot with smoothed line overlaying that represents a non-parametric
fit of a model to the expected outcomes in the population at each quantile
of the joint exposures (e.g. '1' represents 'at the first quantile for
every exposure')
}
\examples{
set.seed(12)
\dontrun{
}
}
\seealso{
\code{\link[qgcomp]{qgcomp.noboot}}, \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp}}
}
