% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{print.qgcompemmfit}
\alias{print.qgcompemmfit}
\title{Default printing method for a qgcompemmfit object}
\usage{
\method{print}{qgcompemmfit}(x, showweights = TRUE, ...)
}
\arguments{
\item{x}{"qgcompemmfit" object from \code{qgcomp.emm.glm.noboot}
function}

\item{showweights}{logical: should weights be printed, if estimated?}

\item{...}{unused}
}
\value{
Invisibly returns x. Called primarily for side effects.
}
\description{
Prints output depending the model used. \code{qgcomp.emm.glm.noboot} and \code{qgcomp.emm.cox.noboot} will output final estimate of joint exposure
effect, as well
as estimates of the 'weights' (scaled coefficients). \code{qgcomp.emm.glm.boot} and \code{qgcomp.emm.glm.ee} methods will only output final effect estimates.
}
\seealso{
\code{\link[qgcompint]{qgcomp.emm.glm.noboot}}, \code{\link[qgcompint]{getstratweights}}
}
\concept{variance mixtures}
