% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_calcweights.R
\name{getstrateffects}
\alias{getstrateffects}
\title{Calculate mixture effect at a set value of effect measure modifier}
\usage{
getstrateffects(x, emmval = 1)
}
\arguments{
\item{x}{"qgcompemmfit" object from qgcomp.emm.noboot
function}

\item{emmval}{numerical: value of effect measure modifier at which weights are generated}

\item{...}{unused}
}
\description{
A standard qgcomp fit with effect measure modification
only estimates effects at the referent (0) level of the modifier (psi1).
This function can be used to estimate effects at arbitrary levels of the modifier
}
\seealso{
\code{\link[qgcompint]{qgcomp.emm.noboot}} \code{\link[qgcompint]{getstratweights}}
}
\concept{variance mixtures}
