% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_strat.R
\name{print.qgcompemmfit}
\alias{print.qgcompemmfit}
\title{Default printing method for a qgcompemmfit object}
\usage{
\method{print}{qgcompemmfit}(x, showweights = TRUE, ...)
}
\arguments{
\item{x}{"qgcompemmfit" object from \code{qgcomp}, \code{qgcomp.noboot} or \code{qgcomp.boot}
function}

\item{showweights}{logical: should weights be printed, if estimated?}

\item{...}{unused}
}
\description{
Gives variable output depending on whether \code{qgcomp.noboot} or \code{qgcomp.boot}
is called. For \code{qgcomp.noboot} will output final estimate of joint exposure
effect (similar to the 'index' effect in weighted quantile sums), as well
as estimates of the 'weights' (standardized coefficients). For \code{qgcomp.boot},
the marginal effect is given, but no weights are reported since this approach
generally incorporates non-linear models with interaction terms among exposures,
which preclude weights with any useful interpretation.
}
\seealso{
\code{\link[qgcomp]{qgcomp.emm.noboot}}, \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp}}
}
\concept{variance mixtures}
